refactor: rename AboutUs component into CompetenciesPage
