+++ src/pages/CompetenciesPage/BottomGridSection/BottomGridSection.module.scss
