+++ src/pages/CompetenciesPage/BottomGridSection/BottomGridSection.tsx
