+++ src/pages/CompetenciesPage/CompetenciesPage.module.scss
