+++ src/pages/CompetenciesPage/CompetenciesPage.tsx
@@ -1,11 +1,11 @@
 import { FC } from 'react';
 import { Outlet, useMatch } from 'react-router-dom';
 import { ROUTE_COMPETENCIES } from '../../constants/constants';
-import s from './AboutUs.module.scss';
 import { BottomGridSection } from './BottomGridSection/BottomGridSection';
+import s from './CompetenciesPage.module.scss';
 import { TopGridSection } from './TopGridSection/TopGridSection';
 
-export const AboutUs: FC = () => {
+export const CompetenciesPage: FC = () => {
   const isCompetenciesDefaultPage = useMatch(ROUTE_COMPETENCIES);
 
   if (isCompetenciesDefaultPage) {
