+++ src/pages/CompetenciesPage/TopGridSection/TopGridSection.module.scss
