+++ src/pages/CompetenciesPage/TopGridSection/TopGridSection.tsx
